clean buadrate settings
