+++ TFT/src/User/API/serialConnection.h
@@ -1,0 +1,21 @@
+#ifndef _SERIAL_CONNECTION_H_
+#define _SERIAL_CONNECTION_H_
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#define ITEM_BAUDRATE_COUNT 9
+
+extern const uint32_t baudrateList[ITEM_BAUDRATE_COUNT];
+
+void Serial_ReSourceDeInit(void);
+void Serial_ReSourceInit(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
