+++ TFT/src/User/Menu/Mode.h
@@ -7,9 +7,6 @@ extern "C" {
 
 void infoMenuSelect(void);
 
-void Serial_ReSourceDeInit(void);
-void Serial_ReSourceInit(void);
-
 #ifdef __cplusplus
 }
 #endif
