+++ TFT/src/User/includes.h
@@ -46,6 +46,7 @@
 #include "HD44780_Simulator.h"
 #include "ui_draw.h"
 #include "touch_process.h"
+#include "serialConnection.h"
 #include "interfaceCmd.h"
 #include "coordinate.h"
 #include "ff.h"
